feat(auth): add signup page
This commit adds signing up capabilities to the application, adding sign
up page and an onboarding setup with 2 possible pages, the login page is
currently not done.

Also rearranges the files in a more organized way where we
have two "modules" called `onboarding/` and `app/` which differentiates
between the onboarding process (signing up and logging in) and the app
itself once you're authenticated.

Also removes the Firebase emulators setup because I had problems with it
running on HTTP and not HTTPS.

Also removes widget test since I don't plan on creating tests anytime
soon and it's outdated.